:recycle: Refactor: Move SignUpType into User Dto
