+++ src/main/java/Honzapda/Honzapda_server/shop/data/entity/Shop.java
@@ -1,6 +1,6 @@
 package Honzapda.Honzapda_server.shop.data.entity;
 
-import Honzapda.Honzapda_server.common.dto.SignUpType;
+import Honzapda.Honzapda_server.user.data.entity.SignUpType;
 import Honzapda.Honzapda_server.user.data.entity.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
