+++ src/main/java/Honzapda/Honzapda_server/user/data/entity/SignUpType.java
@@ -1,4 +1,4 @@
-package Honzapda.Honzapda_server.common.dto;
+package Honzapda.Honzapda_server.user.data.entity;
 
 public enum SignUpType {
         LOCAL, APPLE, GOOGLE, KAKAO
